Brought ButtonGroup in line with Chips
- Now supports rendering individual items as anchors, buttons, OR Links
- Now uses label / key props / fields (content / value still work, but are deprecated)
- Component level utilities have been abstracted to utility files for re-use / clarity / testability
